# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Story Sounds API",
        "description": "REST Apis for Story Sounds",
        "version": "1.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logs a user in",
                "operationId": "login",
                "requestBody": {
                    "description": "User credentials for authentication",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User's email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/auth/social/{provider}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Redirects the user to the Social authentication page",
                "operationId": "redirectToProvider",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The provider login page"
                    }
                }
            }
        },
        "/auth/social/{provider}/callback": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Handle provider callback/Login or Register user",
                "operationId": "handleProviderCallback",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The user details"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Story Sounds authentication endpoints"
        }
    ]
}